Works with floats
Yongming Shen and Prof  Mike Ferdman helped to parse the output from the command line that I was having trouble with. Now able to parse the byte format from the gatttool, splits the line, packs the byte values that we care about together then converts that packed byte into a float. data[0] is used because it is a tuple.